move Datadog.Trace.ClrProfiler.ScopeDBFactory to Datadog.Trace.ClrProfiler.AutoInstrumentation.AdoNet.DbScopeFactory
